added back mutex

+++ dataset/data.py
@@ -74,8 +74,8 @@ class KittiSet(Dataset):
     def __getitem__(self, idx):
         sample = self.samples[idx]
 
-        s = cv.imread(sample["frame"]).astype(np.float32)  #HxWxC
-        s_ = cv.imread(sample["next_frame"]).astype(np.float32)
+        s = cv.imread(sample["frame"])  #HxWxC
+        s_ = cv.imread(sample["next_frame"])
         depth = cv.imread(sample["depth"]).astype(np.float32)
         Rt = sample["poses"]
         k = sample["intrinsic"]
